+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.1.6
+
+- feat: add `ctrl+a` and `ctrl+x` to increase and decrease the first found number in block. Support multiple selections and combo.
+
 ## v0.1.5
 
 - feat: add `:copy-path` to get page or journal absolute path, so you can edit it outside of Logseq.